Заготовили абстрактный класс, вынесли рендер и поправили импорты
