+++ src/components/filters.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import {createElement} from '../utils/render';
 
 const getFilterMarkup = (filterName, isChecked) => {
   return (
