+++ src/components/statistics.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import {createElement} from '../utils/render';
 
 const getStatistics = (statType, statName) => {
