+++ src/components/trip-events.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render';
 
 const getTripEventsTemplate = () => {
   return (
