+++ src/components/trip-header.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render';
 
 const getNewEventBtnTemplate = () => `<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>`;
 const getTripMainTemplate = () => `<div class="trip-main">${getNewEventBtnTemplate()}</div>`;
